Fix autoloading issues.
PHP, by default will look for a class by lowercasing the classname and looking for that file with a .inc or .php extension. On some systems, classes could not be located. This fix will looks for the classname as is (not lowercased) as that is the file structure laid out for this project.